tno change int & TodoDaoService - test

+++ src/main/java/com/practice/todolist/dto/TodoDto.java
@@ -12,7 +12,7 @@ import java.time.LocalDate;
 @NoArgsConstructor
 @AllArgsConstructor
 public class TodoDto {
-    private long tno;
+    private int tno;
     private String title;
     private LocalDate dueDate;
     private boolean finished;
